Significant refactor. Create base classes for PyDreoTowerFan, PyDreoAirCirculator, PyDreoCeilingFan
Purpose of this is to simplify code logic for each variety.
