+++ custom_components/dreo/__init__.py
@@ -7,12 +7,7 @@ from .haimports import *  # pylint: disable=W0401,W0614
 from .const import (
     LOGGER,
     DOMAIN,
-    DREO_FANS,
-    DREO_AIR_PURIFIERS,
-    DREO_HEATERS,
-    DREO_AIRCONDITIONERS,
-    DREO_COOKERS,
-    DREO_MANAGER,
+    PYDREO_MANAGER,
     DREO_PLATFORMS,
     CONF_AUTO_RECONNECT,
 )
@@ -35,64 +30,54 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> b
     region = "us"
 
     from .pydreo import PyDreo  # pylint: disable=C0415
+    from .pydreo.constant import DreoDeviceType
 
-    manager = PyDreo(username, password, region)
-    manager.auto_reconnect = auto_reconnect
+    pydreo_manager = PyDreo(username, password, region)
+    pydreo_manager.auto_reconnect = auto_reconnect
 
-    login = await hass.async_add_executor_job(manager.login)
+    login = await hass.async_add_executor_job(pydreo_manager.login)
 
     if not login:
         _LOGGER.error("Unable to login to the dreo server")
         return False
 
-    load_devices = await hass.async_add_executor_job(manager.load_devices)
+    load_devices = await hass.async_add_executor_job(pydreo_manager.load_devices)
 
     if not load_devices:
         _LOGGER.error("Unable to load devices from the dreo server")
         return False
 
-    device_dict = process_devices(manager)
-    _LOGGER.debug("Device dict is: %s", device_dict)
+    device_types = set()
+    for device in pydreo_manager.devices:
+        device_types.add(device.type)   
 
-    manager.start_transport()
+    _LOGGER.info("%d Dreo devices found", len(pydreo_manager.devices))
 
-    hass.data[DOMAIN] = {}
-    hass.data[DOMAIN][DREO_MANAGER] = manager
-
-    fans = hass.data[DOMAIN][DREO_FANS] = []
-    air_purifiers = hass.data[DOMAIN][DREO_AIR_PURIFIERS] = []
-    heaters = hass.data[DOMAIN][DREO_HEATERS] = []
-    acs = hass.data[DOMAIN][DREO_AIRCONDITIONERS] = []
-    cookers = hass.data[DOMAIN][DREO_COOKERS] = []
     platforms = set()
-
-    if device_dict[DREO_FANS]:
-        fans.extend(device_dict[DREO_FANS])
+    if (DreoDeviceType.TOWER_FAN in device_types or 
+        DreoDeviceType.AIR_CIRCULATOR in device_types or
+        DreoDeviceType.AIR_PURIFIER in device_types):
         platforms.add(Platform.FAN)
         platforms.add(Platform.SENSOR)
         platforms.add(Platform.SWITCH)
         platforms.add(Platform.NUMBER)
 
-    if device_dict[DREO_HEATERS]:
-        heaters.extend(device_dict[DREO_HEATERS])
+    if (DreoDeviceType.HEATER in device_types or 
+        DreoDeviceType.AIR_CONDITIONER in device_types):
         platforms.add(Platform.CLIMATE)
         platforms.add(Platform.SENSOR)
         platforms.add(Platform.SWITCH)
         platforms.add(Platform.NUMBER)
 
-    if device_dict[DREO_AIRCONDITIONERS]:
-        acs.extend(device_dict[DREO_AIRCONDITIONERS])
-        platforms.add(Platform.CLIMATE)
+    if (DreoDeviceType.CHEF_MAKER in device_types):
         platforms.add(Platform.SENSOR)
         platforms.add(Platform.SWITCH)
         platforms.add(Platform.NUMBER)
 
-    if device_dict[DREO_COOKERS]:
-        cookers.extend(device_dict[DREO_COOKERS])
-        platforms.add(Platform.SENSOR)
-        platforms.add(Platform.SWITCH)
-        platforms.add(Platform.NUMBER)
-        
+    pydreo_manager.start_transport()
+
+    hass.data[DOMAIN] = {}
+    hass.data[DOMAIN][PYDREO_MANAGER] = pydreo_manager
     hass.data[DOMAIN][DREO_PLATFORMS] = platforms
 
     _LOGGER.debug("Platforms are: %s", platforms)
@@ -110,7 +95,7 @@ async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> b
 
 async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> bool:
     """Unload a config entry."""
-    manager = hass.data[DOMAIN][DREO_MANAGER]
+    manager = hass.data[DOMAIN][PYDREO_MANAGER]
     if unload_ok := await hass.config_entries.async_unload_platforms(
         config_entry,
         hass.data[DOMAIN][DREO_PLATFORMS],
@@ -119,34 +104,3 @@ async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> 
 
     manager.stop_transport()
     return unload_ok
-
-def process_devices(manager) -> dict:
-    """Assign devices to proper component."""
-    devices = {}
-    devices[DREO_FANS] = []
-    devices[DREO_AIR_PURIFIERS] = []
-    devices[DREO_HEATERS] = []
-    devices[DREO_AIRCONDITIONERS] = []
-    devices[DREO_COOKERS] = []
-
-    if manager.fans:
-        devices[DREO_FANS].extend(manager.fans)
-        _LOGGER.info("%d Dreo fans found", len(manager.fans))
-
-    if manager.air_purifiers:
-        devices[DREO_AIR_PURIFIERS].extend(manager.air_purifiers)
-        _LOGGER.info("%d Dreo air purifiers found", len(manager.air_purifiers))
-
-    if manager.heaters:
-        devices[DREO_HEATERS].extend(manager.heaters)
-        _LOGGER.info("%d Dreo heaters found", len(manager.heaters))
-
-    if manager.acs:
-        devices[DREO_AIRCONDITIONERS].extend(manager.acs)
-        _LOGGER.info("%d Dreo ACs found", len(manager.acs))
-
-    if manager.cookers:
-        devices[DREO_COOKERS].extend(manager.cookers)
-        _LOGGER.info("%d Dreo cookers found", len(manager.cookers))
-
-    return devices
