+++ custom_components/dreo/const.py
@@ -4,14 +4,7 @@ LOGGER = "dreo"
 DOMAIN = "dreo"
 DREO_DISCOVERY = "dreo_discovery_{}"
 SERVICE_UPDATE_DEVS = "update_devices"
-
-DREO_FANS = "fans"
-DREO_AIR_PURIFIERS = "air_purifiers"
-DREO_HEATERS = "heaters"
-DREO_AIRCONDITIONERS = "airconditioners"
-DREO_COOKERS = "cookers"
-DREO_SENSORS = "sensors"
-DREO_MANAGER = "manager"
+PYDREO_MANAGER = "pydreo_manager"
 DREO_PLATFORMS = "platforms"
 
 CONF_AUTO_RECONNECT = "auto_reconnect"
