+++ custom_components/dreo/pydreo/pydreobasedevice.py
@@ -12,6 +12,9 @@ if TYPE_CHECKING:
 
 _LOGGER = logging.getLogger(LOGGER_NAME)
 
+class UnknownProductError(Exception):
+    """Exception thrown when we don't recognize a product of a device."""
+
 class UnknownModelError(Exception):
     """Exception thrown when we don't recognize a model of a device."""
 
@@ -153,6 +156,11 @@ class PyDreoBaseDevice(object):
     def brand(self):
         """Returns the device's manufacturer."""
         return "Dreo"
+
+    @property
+    def type(self):
+        """Returns the device's type."""
+        return self._device_definition.device_type
 
     @property
     def model(self):
