+++ custom_components/dreo/pydreo/pydreoceilingfan.py
@@ -1,0 +1,113 @@
+"""Dreo API for controling fans."""
+
+import logging
+from typing import TYPE_CHECKING, Dict
+
+from .constant import (
+    LOGGER_NAME,
+    FANON_KEY,
+    WINDLEVEL_KEY,
+    SPEED_RANGE,
+    FAN_MODE_STRINGS
+)
+
+from .pydreofan import PyDreoFanBase
+from .models import DreoDeviceDetails
+
+_LOGGER = logging.getLogger(LOGGER_NAME)
+
+if TYPE_CHECKING:
+    from pydreo import PyDreo
+
+
+class PyDreoCeilingFan(PyDreoFanBase):
+    """Base class for Dreo Fan API Calls."""
+
+    def __init__(self, device_definition: DreoDeviceDetails, details: Dict[str, list], dreo: "PyDreo"):
+        """Initialize air devices."""
+        super().__init__(device_definition, details, dreo)
+        
+        self._speed_range = None
+        if (device_definition.speed_range is not None):
+            self._speed_range = device_definition.speed_range[SPEED_RANGE]
+        if (self._speed_range is None):
+            self._speed_range = self.parse_speed_range(details)
+        self._preset_modes = device_definition.preset_modes
+        if (self._preset_modes is None):
+            self._preset_modes = self.parse_preset_modes(details)
+
+        self._fan_speed = None
+
+        self._wind_type = None
+        self._wind_mode = None
+        self._osc_mode = None
+        self._cruise_conf = None
+
+        self._temperature = None
+        self._led_always_on = None
+        self._voice_on = None
+        self._device_definition = device_definition
+        self._light_sensor_on = None
+        self._mute_on = None
+
+        self._fixed_conf = None
+
+    def __repr__(self):
+        # Representation string of object.
+        return f"<{self.__class__.__name__}:{self._device_id}:{self._name}>"
+    
+    def parse_preset_modes(self, details: Dict[str, list]) -> tuple[str, int]:
+        """Parse the preset modes from the details."""
+        preset_modes = []
+        controls_conf = details.get("controlsConf", None)
+        if controls_conf is not None:
+            control = controls_conf.get("control", None)
+            if (control is not None):
+                for control_item in control:
+                    if (control_item.get("type", None) == "CFFan"):
+                        for mode_item in control_item.get("items", None):
+                            text_id = mode_item.get("text", None)
+                            text = FAN_MODE_STRINGS[text_id]
+                            value = mode_item.get("value", None)
+                            preset_modes.append((text, value))
+
+        preset_modes.sort(key=lambda tup: tup[1])  # sorts in place
+        if (preset_modes.count is 0):
+            _LOGGER.debug("PyDreoFan:No preset modes detected")
+            preset_modes = None
+        _LOGGER.debug("PyDreoFan:Detected preset modes - %s", preset_modes)
+        return preset_modes
+
+    @PyDreoFanBase.is_on.setter
+    def is_on(self, value: bool):
+        """Set if the fan is on or off"""
+        _LOGGER.debug("PyDreoFan:is_on.setter - %s", value)
+        self._send_command(FANON_KEY, value)
+
+    @property
+    def oscillating(self) -> bool:
+        return None
+    
+    @oscillating.setter
+    def oscillating(self, value: bool) -> None:
+        raise NotImplementedError(f"Attempting to set oscillating on a device that doesn't support ({value})")
+    
+    def update_state(self, state: dict):
+        """Process the state dictionary from the REST API."""
+        _LOGGER.debug("PyDreoFan:update_state")
+        super().update_state(state)
+
+        self._fan_speed = self.get_state_update_value(state, WINDLEVEL_KEY)
+        if self._fan_speed is None:
+            _LOGGER.error("Unable to get fan speed from state. Check debug logs for more information.")
+
+        self._is_on = self.get_state_update_value(state, FANON_KEY)
+
+    def handle_server_update(self, message):
+        """Process a websocket update"""
+        _LOGGER.debug("PyDreoCeilingFan:handle_server_update")
+        super().handle_server_update(message)
+
+        val_power_on = self.get_server_update_key_value(message, FANON_KEY)
+        if isinstance(val_power_on, bool):
+            self._is_on = val_power_on
