+++ custom_components/dreo/pydreo/pydreotowerfan.py
@@ -1,0 +1,140 @@
+"""Dreo API for controling fans."""
+
+import logging
+from typing import TYPE_CHECKING, Dict
+
+from .constant import (
+    LOGGER_NAME,
+    SHAKEHORIZON_KEY,
+    SHAKEHORIZONANGLE_KEY,
+    SPEED_RANGE,
+    FAN_MODE_STRINGS
+)
+
+from .pydreofan import PyDreoFanBase
+from .models import DreoDeviceDetails
+
+_LOGGER = logging.getLogger(LOGGER_NAME)
+
+if TYPE_CHECKING:
+    from pydreo import PyDreo
+
+
+class PyDreoTowerFan(PyDreoFanBase):
+    """Base class for Dreo Fan API Calls."""
+
+    def __init__(self, device_definition: DreoDeviceDetails, details: Dict[str, list], dreo: "PyDreo"):
+        """Initialize air devices."""
+        super().__init__(device_definition, details, dreo)
+        
+        self._speed_range = None
+        if (device_definition.speed_range is not None):
+            self._speed_range = device_definition.speed_range[SPEED_RANGE]
+        if (self._speed_range is None):
+            self._speed_range = self.parse_speed_range(details)
+        self._preset_modes = device_definition.preset_modes
+        if (self._preset_modes is None):
+            self._preset_modes = self.parse_preset_modes(details)
+
+        self._shakehorizon = None
+        self._shakehorizonangle = None
+
+    def __repr__(self):
+        # Representation string of object.
+        return f"<{self.__class__.__name__}:{self._device_id}:{self._name}>"
+
+    def parse_speed_range_from_control_node(self, control_node) -> tuple[int, int]:
+        """Parse the speed range from a control node"""
+        for control_item in control_node:
+            if control_item.get("type", None) == "Speed":
+                speed_low = control_item.get("items", None)[0].get("value", None)
+                speed_high = control_item.get("items", None)[1].get("value", None)
+                speed_range = (speed_low, speed_high)
+                return speed_range
+        return None
+    
+    def parse_preset_modes(self, details: Dict[str, list]) -> tuple[str, int]:
+        """Parse the preset modes from the details."""
+        preset_modes = []
+        controls_conf = details.get("controlsConf", None)
+        if controls_conf is not None:
+            control = controls_conf.get("control", None)
+            if (control is not None):
+                for control_item in control:
+                    if (control_item.get("type", None) == "Mode" or
+                        control_item.get("type", None) == "CFFan"):
+                        for mode_item in control_item.get("items", None):
+                            text_id = mode_item.get("text", None)
+                            text = FAN_MODE_STRINGS[text_id]
+                            value = mode_item.get("value", None)
+                            preset_modes.append((text, value))
+            schedule = controls_conf.get("schedule", None)
+            if (schedule is not None):
+                modes = schedule.get("modes", None)
+                if (modes is not None):
+                    for mode_item in modes:
+                        text_id = mode_item.get("title", None)
+                        text = FAN_MODE_STRINGS[text_id]
+                        value = mode_item.get("value", None)
+                        if (text, value) not in preset_modes:
+                            preset_modes.append((text, value))
+
+        preset_modes.sort(key=lambda tup: tup[1])  # sorts in place
+        if (preset_modes.count is 0):
+            _LOGGER.debug("PyDreoFan:No preset modes detected")
+            preset_modes = None
+        _LOGGER.debug("PyDreoFan:Detected preset modes - %s", preset_modes)
+        return preset_modes
+
+    @property
+    def oscillating(self) -> bool:
+        """Returns `True` if either horizontal or vertical oscillation is on."""
+        if self._shakehorizon is not None:
+            return self._shakehorizon
+        return None
+
+    @oscillating.setter
+    def oscillating(self, value: bool) -> None:
+
+        """Enable or disable oscillation"""
+        _LOGGER.debug("PyDreoFan:oscillating.setter")
+
+        if self._shakehorizon is not None:
+            self._send_command(SHAKEHORIZON_KEY, value)
+        else:
+            raise NotImplementedError("Attempting to set oscillating on a device that doesn't support.")
+
+    @property
+    def shakehorizonangle(self) -> int:
+        """Get the current oscillation angle"""
+        if self._shakehorizonangle is not None:
+            return self._shakehorizonangle
+
+    @shakehorizonangle.setter
+    def shakehorizonangle(self, value: int) -> None:
+        """Set the oscillation angle."""
+        _LOGGER.debug("PyDreoFan:shakehorizonangle.setter")
+        if self._shakehorizonangle is not None:
+            self._send_command(SHAKEHORIZONANGLE_KEY, value)            
+
+    def update_state(self, state: dict):
+        """Process the state dictionary from the REST API."""
+        _LOGGER.debug("PyDreoFan:update_state")
+        super().update_state(state)
+
+        self._shakehorizon = self.get_state_update_value(state, SHAKEHORIZON_KEY)
+        self._shakehorizonangle = self.get_state_update_value(state, SHAKEHORIZONANGLE_KEY)
+
+    def handle_server_update(self, message):
+        """Process a websocket update"""
+        _LOGGER.debug("PyDreoFan:handle_server_update")
+        super().handle_server_update(message)
+
+        val_shakehorizon = self.get_server_update_key_value(message, SHAKEHORIZON_KEY)
+        if isinstance(val_shakehorizon, bool):
+            self._shakehorizon = val_shakehorizon
+
+        val_shakehorizonangle = self.get_server_update_key_value(message, SHAKEHORIZONANGLE_KEY)
+        if isinstance(val_shakehorizonangle, int):
+            self._shakehorizonangle = val_shakehorizonangle
+
