+++ tests/pydreo/test_pydreoairpurifier.py
@@ -1,11 +1,9 @@
 """Tests for Dreo Fans"""
 # pylint: disable=used-before-assignment
 import logging
-from unittest.mock import patch
-import pytest
 from  .imports import * # pylint: disable=W0401,W0614
 from . import call_json
-from .testbase import TestBase, PATCH_SEND_COMMAND
+from .testbase import TestBase
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
@@ -21,7 +19,7 @@ class TestPyDreoAirPurifier(TestBase):
 
         self.get_devices_file_name = "get_devices_HAP003S.json"
         self.manager.load_devices()
-        assert len(self.manager.air_purifiers) == 1
-        air_purifier = self.manager.air_purifiers[0]
+        assert len(self.manager.devices) == 1
+        air_purifier = self.manager.devices[0]
         assert air_purifier.speed_range == (1, 18)
         assert air_purifier.preset_modes == ['auto', 'manual', 'sleep', 'turbo']
