+++ tests/pydreo/test_pydreoheater.py
@@ -19,8 +19,8 @@ class TestPyDreoHeater(TestBase):
 
         self.get_devices_file_name = "get_device_state_HSH004S_1.json"
         self.manager.load_devices()
-        assert len(self.manager.heaters) == 1
-        heater = self.manager.heaters[0]
+        assert len(self.manager.devices) == 1
+        heater = self.manager.devices[0]
 
         assert heater.heat_range == (1, 3)
         assert heater.preset_modes == ['H1', 'H2', 'H3']
