+++ tests/pydreo/test_pydreotowerfan.py
@@ -20,8 +20,8 @@ class TestPyDreoFan(TestBase):
 
         self.get_devices_file_name = "get_devices_HTF008S.json"
         self.manager.load_devices()
-        assert len(self.manager.fans) == 1
-        fan = self.manager.fans[0]
+        assert len(self.manager.devices) == 1
+        fan = self.manager.devices[0]
         assert fan.speed_range == (1, 5)
         assert fan.preset_modes == ['normal', 'natural', 'sleep', 'auto']
 
@@ -43,74 +43,16 @@ class TestPyDreoFan(TestBase):
         with pytest.raises(ValueError):
             fan.fan_speed = 10
 
-    def test_circulator_load_and_send_commands(self):
-        """Load circulator fan and test sending commands."""
-        self.get_devices_file_name = "get_devices_HAF004S.json"
-        self.manager.load_devices()
-
-        assert len(self.manager.fans) == 1
-
-        fan = self.manager.fans[0]
-
-        assert fan.speed_range == (1, 9)
-        assert fan.preset_modes == ['normal', 'natural', 'sleep', 'auto', 'turbo', 'custom']
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.is_on = True
-            mock_send_command.assert_called_once_with(fan, {POWERON_KEY: True})
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.preset_mode = 'normal'
-            mock_send_command.assert_called_once_with(fan, {WIND_MODE_KEY: 1})
-
-        with pytest.raises(ValueError):
-            fan.preset_mode = 'not_a_mode'
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.fan_speed = 3
-            mock_send_command.assert_called_once_with(fan, {WINDLEVEL_KEY: 3})
-
-        with pytest.raises(ValueError):
-            fan.fan_speed = 10
-
-
-    def test_HCF001S(self):  # pylint: disable=invalid-name
-        """Load fan and test sending commands."""
-
-        self.get_devices_file_name = "get_devices_HCF001S.json"
-        self.manager.load_devices()
-        assert len(self.manager.fans) == 1
-        fan = self.manager.fans[0]
-        assert fan.speed_range == (1, 12)
-        assert fan.preset_modes == ['normal', 'natural', 'sleep', 'reverse']
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.is_on = True
-            mock_send_command.assert_called_once_with(fan, {POWERON_KEY: True})
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.preset_mode = 'normal'
-            mock_send_command.assert_called_once_with(fan, {MODE_KEY: 1})
-
-        with pytest.raises(ValueError):
-            fan.preset_mode = 'not_a_mode'
-
-        with patch(PATCH_SEND_COMMAND) as mock_send_command:
-            fan.fan_speed = 3
-            mock_send_command.assert_called_once_with(fan, {WINDLEVEL_KEY: 3})
-
-        with pytest.raises(ValueError):
-            fan.fan_speed = 13
-
     def test_HTF005S(self):  # pylint: disable=invalid-name
         """Load fan and test sending commands."""
 
         self.get_devices_file_name = "get_devices_HTF005S.json"
         self.manager.load_devices()
-        assert len(self.manager.fans) == 1
-        fan = self.manager.fans[0]
+        assert len(self.manager.devices) == 1
+        fan = self.manager.devices[0]
         assert fan.speed_range == (1, 12)
         assert fan.preset_modes == ['normal', 'natural', 'sleep', 'auto']
+        assert fan.oscillating is True
 
         with patch(PATCH_SEND_COMMAND) as mock_send_command:
             fan.is_on = True
